move function to map UPRNs to clusters to cluster.py
